@@ -36,8 +36,7 @@ classifiers =
     Framework :: tox
     Framework :: Sphinx
     Programming Language :: Python
-    Programming Language :: Python :: 3.10
-    Programming Language :: Python :: 3.11
+    Programming Language :: Python :: 3.13
     Programming Language :: Python :: 3 :: Only
 keywords =
     snekpack
@@ -60,7 +59,7 @@ install_requires =
 # Random options
 zip_safe = false
 include_package_data = True
-python_requires = >=3.8
+python_requires = >=3.13
 
 # Where is my code
 packages = find:
